Title: 정수 명령 처리 2, Time: 243ms, Memory: 12MB, Status: Passed - Codetree

+++ 250929/정수 명령 처리 2/process-numeric-commands-2.java
@@ -13,7 +13,7 @@ public class Main {
             String order = st.nextToken();
             if(order.equals("push")) {
                 int tmp = Integer.parseInt(st.nextToken());
-                q.add((Integer)tmp);
+                q.add(tmp);
                 continue;
             } else if(order.equals("front")){
                 System.out.println(q.peek());
